+++ CHANGELOG.md
@@ -6,11 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+- Add getActiveOrganizationsByEmail to return only active organizations
 
 ## [0.61.1] - 2024-10-29
 
 ### Fixed
 - Avoid calls to checkUserPermissions when session data is not available
+
 
 ## [0.61.0] - 2024-10-16